add memory profiling arg test
