+++ tests/profiler/test_profiler.py
@@ -1,12 +1,19 @@
 # Copyright 2022 MosaicML Composer authors
 # SPDX-License-Identifier: Apache-2.0
 
+from typing import Union
 from unittest.mock import MagicMock
 
 import pytest
 
 from composer.core import State
 from composer.profiler import Profiler, ProfilerAction, SystemProfiler, TorchProfiler, cyclic_schedule
+from packaging import version
+import os
+import pathlib
+import torch
+
+from composer.profiler.utils import export_memory_timeline_html
 
 
 @pytest.mark.parametrize('repeat', [1, 0])
@@ -50,6 +57,7 @@ def test_profiler_init(minimal_state: State):
         trace_handlers=[mock_trace_handler],
         schedule=cyclic_schedule(),
         torch_prof_profile_memory=True,
+        torch_prof_memory_filename=None,
         sys_prof_cpu=True,
     )
     profiler.bind_to_state(minimal_state)
@@ -59,10 +67,9 @@ def test_profiler_init(minimal_state: State):
 
 def test_marker(dummy_state: State):
     mock_trace_handler = MagicMock()
-    profiler = Profiler(
-        trace_handlers=[mock_trace_handler],
-        schedule=cyclic_schedule(),
-    )
+    profiler = Profiler(trace_handlers=[mock_trace_handler],
+                        schedule=cyclic_schedule(),
+                        torch_prof_memory_filename=None)
     profiler.bind_to_state(dummy_state)
     dummy_state.profiler = profiler
     marker = profiler.marker('name',
@@ -94,3 +101,73 @@ def test_marker(dummy_state: State):
 
     assert mock_trace_handler.process_duration_event.call_count == 8
     assert mock_trace_handler.process_instant_event.call_count == 1
+
+
+@pytest.mark.parametrize('torch_prof_with_stack', [True, False])
+@pytest.mark.parametrize('torch_prof_record_shapes', [True, False])
+@pytest.mark.parametrize('torch_prof_profile_memory', [True, False])
+@pytest.mark.parametrize('torch_prof_memory_filename', [None, 'test.html'])
+def test_profiler_error_message(torch_prof_with_stack: bool, torch_prof_record_shapes: bool,
+                                torch_prof_profile_memory: bool, torch_prof_memory_filename: Union[None, str]) -> None:
+    # Construct a profiler and assert that it triggers the ValueError if the arguments are invalid
+    if (torch_prof_memory_filename is not None and
+            not (torch_prof_with_stack and torch_prof_record_shapes and torch_prof_profile_memory)):
+        with pytest.raises(ValueError):
+            _ = Profiler(
+                trace_handlers=[MagicMock()],
+                schedule=cyclic_schedule(),
+                torch_prof_with_stack=torch_prof_with_stack,
+                torch_prof_record_shapes=torch_prof_record_shapes,
+                torch_prof_profile_memory=torch_prof_profile_memory,
+                torch_prof_memory_filename=torch_prof_memory_filename,
+            )
+    else:
+        _ = Profiler(
+            trace_handlers=[MagicMock()],
+            schedule=cyclic_schedule(),
+            torch_prof_with_stack=torch_prof_with_stack,
+            torch_prof_record_shapes=torch_prof_record_shapes,
+            torch_prof_profile_memory=torch_prof_profile_memory,
+            torch_prof_memory_filename=torch_prof_memory_filename,
+        )
+
+
+@pytest.mark.gpu
+def test_memory_timeline(tmp_path: pathlib.Path) -> None:
+    if version.parse(torch.__version__) <= version.parse('2.1.0.dev'):
+        # memory timeline is supported after PyTorch 2.1.0.
+        return
+    import torch.profiler._memory_profiler as _memory_profiler
+
+    model = torch.nn.Sequential(
+        torch.nn.Linear(1024, 1024, bias=True),
+        torch.nn.ReLU(),
+        torch.nn.Linear(1024, 1024, bias=False),
+        torch.nn.Softmax(dim=1),
+    ).to('cuda')
+    optimizer = torch.optim.Adam(model.parameters(), lr=0.1)
+
+    x = torch.ones((1024, 1024), device='cuda')
+    targets = torch.ones((1024, 1024), device='cuda')
+    with torch.profiler.profile(record_shapes=True, with_stack=True, profile_memory=True) as prof:
+        y = model(x)
+        loss = torch.nn.functional.mse_loss(y, targets)
+        loss.backward()
+        optimizer.step()
+        optimizer.zero_grad()
+
+    memory_profile = prof._memory_profile()
+    timeline = memory_profile.timeline
+
+    # this checks the default memory timeline event value (t == -1) for preexisting tensors
+    assert all((t == -1) if action == _memory_profiler.Action.PREEXISTING else (t > 0) for t, action, _, _ in timeline)
+
+    fig = export_memory_timeline_html(
+        prof,
+        os.path.join(tmp_path, 'test_memory_timeline.html'),
+        yxis_step_size=0.01,
+        return_fig=True,
+    )
+    assert fig is not None, 'export_memory_timeline_html should return a figure when return_fig=True'
+    _, end = fig.gca().get_ylim()
+    assert round(end, 2) == 0.06
